Rename endpoint_props module to endpoint_info
