+++ src/mytardis_client/endpoint_info.py
